Level 2: Simple validation
For this level, our server must:

- reject requests to create messages if the message objects have an empty or missing text or from property.

In this case our server should return a status code of 400.

(Advanced note: people don't actually agree on the best status code for this situation.)

A note on security
There is intentionally no security or ownership of messages - anyone can delete one or all messages on our server.

This is a big topic for further study. We won't try to cover it in this challenge.